// fonted/src/app/model/property/property.ts
import {IpropertyBase} from './../propertyBase/propertyBase';
export class Property implements IpropertyBase{
  Id?: string | number ;
  SellRent?: string | number ;
  BHK?: string | number ;
  PType?: string ;
  fType?: string ;
  Name?: string ;
  city?: string ;
  Price?: string ;
  BuildArea?: string | number ;
  security?: string ;
  maintance?: string ;
  carpetArea?: string | number ;
  address?: String ;
  floor?: string | number ;
  totalFloor?: string | number ;
  landMark?: string ;
  RTM?: string | number ;
  date?: string | number ;
  age?: string | number ;
  community?: string ;
  entrance?: string ;
  description?: string ;
  Image?: string ;
  postOn?: string ;
  postBy?: string ;


}